Verified the Blob Mocving

+++ FuncApp-TextExtractor/FuncApp-TextExtractor/ImagesStorage/ImagesStorageService.cs
@@ -1,12 +1,14 @@
 ﻿using Azure.Storage.Blobs;
 using FuncApp_TextExtractor.Configuration;
+using Microsoft.Extensions.Logging;
 using Microsoft.Extensions.Options;
 
 namespace FuncApp_TextExtractor.ImagesBlobStorage;
 
-public class ImagesBlobStorageService(IOptions<FunctionSettings> settings) : IImagesBlobStorageService
+public class ImagesStorageService(IOptions<FunctionSettings> settings, ILogger<ImagesStorageService> logger) : IImagesStorageService
 {
     private readonly FunctionSettings _settings = settings.Value ?? throw new ArgumentNullException(nameof(settings));
+    private readonly ILogger<ImagesStorageService> _logger = logger ?? throw new ArgumentNullException(nameof(logger));
 
     public async Task MoveImageToProcessedContainerAsync(string imageName)
     {
@@ -26,13 +28,24 @@ public class ImagesBlobStorageService(IOptions<FunctionSettings> settings) : IIm
         // Get a reference to the source blob
         var sourceBlobClient = sourceContainerClient.GetBlobClient(imageName);
 
-        // Get a reference to the destination blob
-        var destinationBlobClient = destinationContainerClient.GetBlobClient(imageName);
+        // Check if the source blob exists
+        if (await sourceBlobClient.ExistsAsync())
+        {
+            // Get a reference to the destination blob
+            var destinationBlobClient = destinationContainerClient.GetBlobClient(imageName);
 
-        // Copy the blob from the source to the destination
-        await destinationBlobClient.StartCopyFromUriAsync(sourceBlobClient.Uri);
+            // Copy the blob from the source to the destination
+            await destinationBlobClient.StartCopyFromUriAsync(sourceBlobClient.Uri);
 
-        // Delete the blob from the source container
-        await sourceBlobClient.DeleteIfExistsAsync();
+            // Delete the blob from the source container
+            await sourceBlobClient.DeleteIfExistsAsync();
+        }
+        else
+        {
+            // Handle the case where the source blob does not exist
+            // For example, log a message or throw an exception
+            //throw new FileNotFoundException("Source blob does not exist.");
+            _logger.LogError($"Source blob {imageName} does not exist.");
+        }
     }
 }
